Create a popup window for when the game is over
A popup window is created when the game is over. The content of the
window depends on how the player has done, i.e if the course is completed or
not. Menu and Quit buttons are available in both cases and
Omregistrering is available when the player has failed the course.