Add scoring system based on grades
Add a scoring system based on grades from A to F. When a player is
colliding with a Plutten they get komplettering and their grade is
decreased by one point.

Also included is a timeout of komplettering of 1 second, as in: you
can't get more than one komplettering per second.

The player has lost the game when they reach F. They win the game if
they manage to stay above F for the entire duration of the course.